pyproject.toml
[project]
name = "screen-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deep-translator>=1.11.4",
    "grim>=0.1.10",
    "keyboard>=0.13.5",
    "mss>=10.0.0",
    "pillow>=11.3.0",
    "pyautogui>=0.9.54",
    "pynput>=1.8.1",
    "pytesseract>=0.3.13",
    "ruff>=0.12.2",
]
